change int to long

+++ src/main/java/com/example/calculator/CalculatorApplication.java
@@ -8,7 +8,7 @@ import	org.springframework.stereotype.Service;
 @Service
 public class CalculatorApplication {
 
-	int sum(int a, int b){
+	long sum(long a, long b){
 		return (a+b);
 	}
