+++ src/main/java/com/example/calculator/CalculatorApplicationController.java
@@ -11,7 +11,7 @@ public class CalculatorApplicationController {
     private CalculatorApplicationController calculatorApplication;
 
     @RequestMapping("/sum")
-    String sum (@RequestParam("a") Integer a, @RequestParam("b") Integer b) {
+    String sum (@RequestParam("a") Long a, @RequestParam("b") Long b) {
         return String.valueOf(calculatorApplication.sum(a,b));
     }
 }
